requirements.txt
numpy==1.24.1
onnxruntime==1.19.0
opencv-python==4.10.0.84
pillow==10.2.0
pygame>=2.0.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
uniface==0.1.6
pyautogui>=0.9.54
scipy>=1.10.0
pyserial>=3.5
